Use Poetry for version handling
Getting the version of the project can be done with
the "poetry version" (-s for only the number without name).
Bumping the version can be done also with the same command
but by giving it a bump rule, for example:
"poetry version patch".
Makefile has been updated to call the new commands.