Use Poetry for everything, deleted setup.py
Poetry can do everything setup.py could do,
Makefile has been modified to use poetry commands
instead of pip, virtualenv or pip-tools.

The only difference is that poetry only uses ".venv" for the
in-project virtualenv, which is an advantage IMO, because lots
of tools can pick it up and use that automatically (e.g. VS Code),
but this can be made configurable if necessary:
https://python-poetry.org/docs/configuration/