Parametrize (generalize) simple file content checks
The quilt_project_path test fixture did not do anything quilt-
specific, so pass in the project name through the request parameter to
make it (almost) reusable, and rename it to git_import, because that's
all it does.  Give the internal variables more generic names as well.

Consequently, parametrize the test with the project name and the file
data to check, and request the git_import fixture passing through the
former.

Finally, rename the test source the reflect its widened applicability.