bldr/bldr.py: set USER env var inside the container
Some packages and scripts may query the user name from the USER
environment variable.

One of libssh2 tests has failed because of the unset USER variable.
It tried to use a defaul username the 'username', but the container has
not got this user and other systems may has not got it, too.