bldr/data/hooks/pre-install-deps: add unminimization
Performing an image unminimization, because we experienced that more and more
packages have tests depending on data which is not present on a minimized
image.

For example some affected packages: libpsl, perl, vim, etc...

The ubuntu's image minimaztion method does not only ignore packages, it
also filters files are partof packages. After a package install the filtered
files are not going to present on the filesystem.
You can see the excluding patterns in a minimized image in
/etc/dpkg/dpkg.cfg.d/excludes.