bldr/data/Dockerfile.tpl: move fake local_apt repo creation earlier
The `apt update` command did not work in the `pre-init` hook because
the `local_apt` repository did not exist yet. Only the
'init-docker-image' script created a empty repository (which is
overriden by a bind mount when the contaier is running) to make
`apt update` work at image build time. But `init-docker-image` was
called after `pre-init`, so we had to move the empty repo creation
before calling the `pre-init` hook in the Dockerfile.tpl.