bldr/docker_utils.py: increase timeout of docker API calls
In the communication of docker socket if there is no data transfer for 60
seconds then the docker API hides a timeout exception. But it caused that
the output stream of exec is stopped before the command running finished.
The exec inspect cannot determine the exit code of the running command and
tries to compare with None, which cause the following exception:

TypeError: '>' not supported between instances of 'NoneType' and 'int'