bldr/docker_utils.py: make the docker client usage explicit
There was some shortcuts for user convenience in the docker client
initialization, but it could be misleading and hard to use the same docker
client everywhere. In the following commit we want to pass a parameter to
the create_client function, which was implicit in some cases.

Make the logger initialization in unified form.